When the user login as a guest he can see the brand items

+++ app/src/main/java/com/example/eshfeenygraduationproject/eshfeeny/productsAdapter/BrandItemsAdapter.kt
@@ -19,9 +19,10 @@ import com.varunest.sparkbutton.SparkEventListener
 
 class BrandItemsAdapter(
     private val viewModel: ProductViewModel,
-    val userId: String,
-    val favoriteProducts: ProductResponse,
-    val cartProducts: CartResponse
+    val userId: String?,
+    val favoriteProducts: ProductResponse?,
+    val cartProducts: CartResponse?,
+    val state: String
 ) : ListAdapter<ProductResponseItem, BrandItemsAdapter.ViewHolder>(CategoryDiffCallback()) {
 
     override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): ViewHolder {
@@ -45,11 +46,13 @@ class BrandItemsAdapter(
         fun bind(product: ProductResponseItem) {
 
             setData2UI(product)
-            addItemToCart(product)
-            setFavoriteItem(product)
             navigate2Details(product)
-            incrementProductAmount(product)
-            decrementProductAmount(product)
+            if (state != "guest") {
+                addItemToCart(product)
+                setFavoriteItem(product)
+                incrementProductAmount(product)
+                decrementProductAmount(product)
+            }
         }
 
         private fun setData2UI(product: ProductResponseItem) {
@@ -69,7 +72,7 @@ class BrandItemsAdapter(
         }
 
         private fun setFavoriteItem(category: ProductResponseItem) {
-            if (favoriteProducts.contains(category)) {
+            if (favoriteProducts?.contains(category)!!) {
                 itemBinding.heartIconId.isChecked = true
             }
 
@@ -77,13 +80,17 @@ class BrandItemsAdapter(
                 override fun onEvent(button: ImageView, buttonState: Boolean) {
                     if (buttonState) {
                         // Button is
-                        viewModel.addMedicineToFavorites(
-                            userId,
-                            PatchString(category._id)
-                        )
+                        if (userId != null) {
+                            viewModel.addMedicineToFavorites(
+                                userId,
+                                PatchString(category._id)
+                            )
+                        }
                     } else {
                         // Button is inactive
-                        viewModel.deleteFavoriteProduct(userId, category._id)
+                        if (userId != null) {
+                            viewModel.deleteFavoriteProduct(userId, category._id)
+                        }
                     }
                 }
 
@@ -101,7 +108,7 @@ class BrandItemsAdapter(
 
             itemBinding.add2CartBtn.setOnClickListener {
 
-                itemCount = getQuantityInCart(cartProducts, product._id)
+                itemCount = getQuantityInCart(cartProducts!!, product._id)
 
                 itemBinding.add2CartBtn.visibility = View.GONE
                 itemBinding.cardFunctionalityLayout.visibility = View.VISIBLE
@@ -110,7 +117,9 @@ class BrandItemsAdapter(
                     itemBinding.productAmount.text = itemCount.toString()
                 } else {
                     itemCount++
-                    viewModel.addProductToCart(userId, PatchString(product._id))
+                    if (userId != null) {
+                        viewModel.addProductToCart(userId, PatchString(product._id))
+                    }
                     itemBinding.productAmount.text = itemCount.toString()
                 }
             }
@@ -130,7 +139,9 @@ class BrandItemsAdapter(
 
                 itemCount++
                 itemBinding.productAmount.text = itemCount.toString()
-                viewModel.incrementProductNumberInCart(userId, product._id)
+                if (userId != null) {
+                    viewModel.incrementProductNumberInCart(userId, product._id)
+                }
             }
         }
 
@@ -138,13 +149,16 @@ class BrandItemsAdapter(
             itemBinding.decreaseBtnId.setOnClickListener {
                 if (itemCount == 1) {
 
-                    viewModel.removeProductFromCart(userId, product._id)
+                    if (userId != null) {
+                        viewModel.removeProductFromCart(userId, product._id)
+                    }
                     itemCount--
-
                     itemBinding.add2CartBtn.visibility = View.VISIBLE
                     itemBinding.cardFunctionalityLayout.visibility = View.GONE
                 } else {
-                    viewModel.decrementProductNumberInCart(userId, product._id)
+                    if (userId != null) {
+                        viewModel.decrementProductNumberInCart(userId, product._id)
+                    }
                     itemCount--
                     itemBinding.productAmount.text = itemCount.toString()
                 }
